Separate log entry preparation and formatting from logging.
This makes it easier to extend the logging by patching those methods.

The method ``prepare_log_entry`` prepares a dict with information for
the log entry, which is then passed to the ``format_log_entry`` for
formatting a message-string.
The ``log`` method only calls those two methods and then calls the
logger.

By separating those steps we can easily extend logging data by patching
``format_log_entry``.
Existing parts of the log message can be changed by patching
``prepare_log_entry`` and modifiying the result of a super-call to the
original method.